requirements.txt
opencv-python
configparser
pycryptodome